init again on preseneion or main (comintion option enabled)
enabled combination this veriosn

now i made here a big update on the softeware with mysql sub , to interact with createion and detroying the data and updating and searching on simple strcut of quaries type string in java as sub --0 equ. , and created on the extra implemtnion of the classes on the usage or prob. of the usage and also created on the multi-dupplicated tables on the same databases to avoid any minipulation by human or misleading from machine , there is multiple of fucntions that track on the child tree ,thats works separtly functions and i created the multiple in case the user requireemnts on it in future or in the base also , on reports there is one direction on the usage there is no report deletion or editing it's not enabled in my system created , to avoid also the minipulation , i interacted with txt files before any uploading any thing of my busssines inside the dbs created , now i think the desgin is established and decleared and also finsihed ! , from now i will create direct of the gui and implement the text data, any creation of the sub it's will be just a recall of the class,my english is not so good and i will not fix it also, i will create the structure if you want to read the code and understand and make effort to learn that good not bad , thats will make your vision more better, but at the end my character will never be changed over my variable of the envirmeomnt , i can change the envirment but the envirmetn can't, you can say that wise stament inserted today !! (^~^)